fix memory alloc issue found during fuzzing
This issue caused an occasional crasher case during fuzzing. The
change prevents memory allocation attacks, which could occur
when the parsed message indicates a push data size which is
out of range for the parsed message.

No maximum limit on extracted chunk data size is needed since
in practice the OP_RETURN maximum size will always be
relatively small.

I also renamed the variable "_len" to "chunkLen" since having a
leading "_" in variable names is not used in golang world.